Separate DataType code
